Refactor the go_binding to binding
