+++ internal/binding/rag_analyzer.go
